refactor: move DonationInfo to it's own component
